Reduce DPDK build time in DPDK testing action
Our github actions were being terminated because DPDK 22.07
takes over 20 minutes to build (!), so I've disabled the
building of almost all of the drivers.

Also: force libfl-dev to be installed on the container for
the DPDK test builds to try and resolve problems where
Ubuntu 22.04 builds fail because flex is not being found.